requirements.txt
Flask-JSON >= 0.3.4
PyJWT >= 2.2.0
nodejs >= 0.1.1
jwt>=1.3.1
psycopg2-binary>=2.9.1
bcrypt>=3.2.0
